rename libcephfssqlite.h to libcephfsfsqlite.cc
